requirement.txt
async-timeout==4.0.2
autopep8==1.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
idna==3.3
packaging==21.3
pycodestyle==2.8.0
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
spotipy==2.20.0
toml==0.10.2
urllib3==1.26.9
wrapt==1.14.1
